Create heatmaps_sorted_by_peaks.m
This MATLAB script will take a set of matrices (as the ones generated by BED2matrix.sh) and will plot a heatmap for each file. It will sort the histones in descending order according to the heights of the peaks, provided in a separated file. It will also plot a heatmap with the number of Reads-in-Peaks (RiP) from the peaks file and will add an empty row in the histone profile if it doesn't have nothing at this genomic region.